pyproject.toml
[tool.poetry]
name = "ad-profit-task"
version = "0.1.0"
description = ""
authors = ["C-Hurtmann <constantine2903@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
d = "^0.2.2"
psycopg2-binary = "^2.9.8"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
